Build 3725: three small changes:  * Inhibits use of the default 'openvpn' domain if the per-configuration '-doNotUseDefaultDomain' preference is set.
 * Adds detailed logging of 'Set name server' activities if the 'DB-ALL' or 'DB-UP' preferences are set.

 * Allows use of 'Set nameserver' in debugging versions of Tunnelblick (disabled only by previous build).

+++ tunnelblick/client.route-pre-down.tunnelblick.sh
@@ -51,6 +51,7 @@ EOF
 
 ARG_MONITOR_NETWORK_CONFIGURATION="$(echo "${TUNNELBLICK_CONFIG}" | grep -i '^[[:space:]]*MonitorNetwork :' | sed -e 's/^.*: //g')"
 LEASEWATCHER_PLIST_PATH="$(echo "${TUNNELBLICK_CONFIG}" | grep -i '^[[:space:]]*LeaseWatcherPlistPath :' | sed -e 's/^.*: //g')"
+REMOVE_LEASEWATCHER_PLIST="$(echo "${TUNNELBLICK_CONFIG}" | grep -i '^[[:space:]]*RemoveLeaseWatcherPlist :' | sed -e 's/^.*: //g')"
 PSID="$(echo "${TUNNELBLICK_CONFIG}" | grep -i '^[[:space:]]*Service :' | sed -e 's/^.*: //g')"
 SCRIPT_LOG_FILE="$(echo "${TUNNELBLICK_CONFIG}" | grep -i '^[[:space:]]*ScriptLogFile :' | sed -e 's/^.*: //g')"
 # Don't need: ARG_RESTORE_ON_DNS_RESET="$(echo "${TUNNELBLICK_CONFIG}" | grep -i '^[[:space:]]*RestoreOnDNSReset :' | sed -e 's/^.*: //g')"
@@ -79,7 +80,9 @@ grep Service | sed -e 's/.*Service : //'
         # Remove leasewatcher
         if ${ARG_MONITOR_NETWORK_CONFIGURATION} ; then
             launchctl unload "${LEASEWATCHER_PLIST_PATH}"
-            rm -f "${LEASEWATCHER_PLIST_PATH}"
+            if ${REMOVE_LEASEWATCHER_PLIST} ; then
+                rm -f "${LEASEWATCHER_PLIST_PATH}"
+            fi
             logMessage "Cancelled monitoring of system configuration changes"
             
             # Indicate leasewatcher has been removed
